Currently, if a prerequisite for a build is missing, then the build fails. This experience can be improved if the missing prerequisite is acquired, and made available for the build to consume. Summary of changes: 1. Metadata of each required build tool will be in `.toolversions` file located in repository root 2. Native build invokes a probing mechanism to get a required tool. Build should specify the name of the required tool. Optionally build can also specify if the declared version, which the version that has been verified to work correctly with the product, of the tool is required. Probing mechanism invokes search and acquire scripts to get the tool on behalf of the build, and returns the tool path if successful or an error message. 3. Scripts that search and acquire a tool will be in `tools-local` folder in repository root. Short description of the design is provided below. Implementation of this design for probing CMake is ready.
The design of these scripts is like having ISearchTool and IAcquireTool interfaces. ISearchTool provides a base implementation of searching for the specified tool in environment path, and the search paths specified in the tools metadata file. Base implementation is in `tools-local/unix/search-tool.sh`, and a corresponding PowerShell script for Windows. A tool can override this implementation if it wants to follow a customized search process. This override script will go into the tool specific folder. For example, if CMake overrides the base implementation then, the override script should be in `tools-local/unix/cmake/search-tool.sh`.

ISearchTool has abstract methods that each tool must provide an implementation. For example, `get-version.sh` gets the version of CMake at the specified path. This script will be in tool specific folder at `tools-local/unix/cmake/get-version.sh`.

IAcquireTool provides similar virtual and abstract functions for acquiring a tool.
Entry point to these scripts is `probe-tool.sh` that gets invoked by `src/Native/build-native.sh`.

In the current implementation, a tool is downloaded directly from the internet. If a developer would like to acquire the tool from a different location, then the developer can either modify `tools-local/unix/acquire-tool.sh` within the local repository or have the customized script in a folder rooted outside the repository.  Path to the folder containing such customized scripts can be specified using `OverrideScriptsFolderPath` argument. A future addition would be to provide an option to override the default `.toolversions` file.

+++ src/Native/build-native.sh
@@ -64,8 +64,18 @@ check_native_prereqs()
 {
     echo "Checking pre-requisites..."
 
-    # Check presence of CMake on the path
-    hash cmake 2>/dev/null || { echo >&2 "Please install cmake before running this script"; exit 1; }
+    # Check for CMake
+    ProbeValue=$("$__rootRepo/tools-local/unix/probe-tool.sh" cmake "$__rootRepo" "$__OverrideScriptsFolderPath" "$__StrictToolVersionMatch")
+
+    if [ $? -ne 0 ] || [ ! -f "$ProbeValue" ]; then
+        echo "$ProbeValue"
+        exit 1
+    fi
+
+    # Update environment path to include the path to CMake that the build should consume.
+    CMakeExecutableFolderPath=$(cd "$(dirname "$ProbeValue")"; pwd -P)
+    export PATH="$PATH:$CMakeExecutableFolderPath"
+    echo "CMakePath=$ProbeValue"
 
     # Check for clang
     hash clang-$__ClangMajorVersion.$__ClangMinorVersion 2>/dev/null ||  hash clang$__ClangMajorVersion$__ClangMinorVersion 2>/dev/null ||  hash clang 2>/dev/null || { echo >&2 "Please install clang before running this script"; exit 1; }
@@ -146,6 +156,8 @@ __VerboseBuild=false
 __ClangMajorVersion=0
 __ClangMinorVersion=0
 __StaticLibLink=0
+__StrictToolVersionMatch=""
+__OverrideScriptsFolderPath=""
 __PortableBuild=0
 
 CPUName=$(uname -p)
@@ -245,6 +257,13 @@ while :; do
             ;;
         stripsymbols)
             __CMakeExtraArgs="$__CMakeExtraArgs -DSTRIP_SYMBOLS=true"
+            ;;
+        stricttoolversionmatch)
+            __StrictToolVersionMatch="strict"
+            ;;
+        --overridescriptsfolderpath)
+            shift
+            __OverrideScriptsFolderPath="$1"
             ;;
         --targetgroup)
             shift
